requirements.txt
discord.py
python-dotenv
apscheduler
requests
nltk
spacy
sqlalchemy
asyncpg
google-generativeai
